backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
lightrag-hku==1.4.9.11
PyMuPDF==1.24.14
python-docx==1.1.2
openpyxl==3.1.5
pytesseract==0.3.13
Pillow==11.0.0
openai==1.54.0
httpx==0.27.2
pyinstaller==6.11.1
